pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scoring_engine"
version = "2.0.0"
description = "Scoring Engine for Red/White/Blue Team Competitions"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.10"
authors = [
  {name = "Brandon Myers", email = "pwnbus@mozilla.com"},
  {name = "Rusty Bower", email = "rusty@rustybower.com"},
  {name = "Zack Allen", email = "zallen@fastly.com"},
]

dependencies = [
  "bcrypt==5.0.0",
  "celery[redis]==5.6.2",
  "configparser==7.2.0",
  "cryptography==46.0.5",
  "Flask==3.1.3",
  "Flask-Caching==2.3.1",
  "Flask-Login==0.6.3",
  "Flask-SQLAlchemy==3.1.1",
  "Flask-WTF==1.2.2",
  "mysqlclient==2.2.8",
  "python-dateutil==2.9.0.post0",
  "pytz==2026.1.post1",
  "PyYAML==6.0.3",
  "Werkzeug==3.1.5",
  "uWSGI==2.0.31",
  "mammoth==1.9.0",
  "odfpy==1.4.1",
]

[project.urls]
Homepage = "https://github.com/scoringengine/scoringengine"
Download = "https://github.com/scoringengine/scoringengine/archive/master.zip"

[tool.setuptools.packages.find]
include = ["scoring_engine*" ]

[tool.bumpversion]
current_version = "2.0.0"
commit = true
tag = true
tag_name = "v{new_version}"
message = "Bump version: {current_version} → {new_version}"

[[tool.bumpversion.files]]
filename = "pyproject.toml"
search = 'version = "{current_version}"'
replace = 'version = "{new_version}"'

[[tool.bumpversion.files]]
filename = "scoring_engine/version.py"
search = 'BASE_VERSION = "{current_version}"'
replace = 'BASE_VERSION = "{new_version}"'

[tool.pytest.ini_options]
filterwarnings = [
    # flask-caching uses deprecated init internally
    "ignore::DeprecationWarning:flask_caching",
    # flask-login 0.6.x uses datetime.utcnow() — fixed in 1.0+
    "ignore:datetime.datetime.utcnow:DeprecationWarning:flask_login",
]

[tool.flake8]
exclude = [".git", "*__init__.py"]
ignore = ["E501"]

[tool.isort]
profile = "black"
line_length = 120

[tool.black]
line-length = 120
